--- agent/yubiagent/testdata/Unittest_Attestation_f9.crt
@@ -1,25 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIELTCCAhWgAwIBAgIRAIle8roTGm+E8Kegjx0kv8UwDQYJKoZIhvcNAQELBQAw
-GzEZMBcGA1UEAxMQVW5pdHRlc3QgUm9vdCBDQTAgFw0yMjA4MjkxNzE4MzZaGA8y
-MTAyMDgyOTE3MTgzNlowIjEgMB4GA1UEAxMXVW5pdHRlc3QgQXR0ZXN0YXRpb24g
-ZjkwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQDah3gEUaAAPJLOvL30
-dMMgcsOjC5SVZDb2ayDuo6Ce4rMCXJ0LUl3lmb760X0pgaXD5Dt+4/34C8ViOQaC
-WSfiQPFYUKc+9NVp9XXodCw6z7lrj3gt8N7SGNRUd6J2ivihyusD3y1S8r0NukbE
-WQJykiQwV7xW28lRDsM5lc7rgqVaVT/NbJBL5wrWpJBnfltXKR4WCUMCfMRyVd/1
-igzcE7zeTqN3RNwZn+XLjYNw01B52n+Pv8e99NjvHYbHb276O2OmKHldiJmpXsKu
-SFIVc7uezzNGCa+27uAqiuBFh8cseWekSr8hb9J//JERx2rLfjF1DOfJJZAtBFlr
-7V9/AgMBAAGjYzBhMA4GA1UdDwEB/wQEAwIBBjAPBgNVHRMBAf8EBTADAQH/MB0G
-A1UdDgQWBBTp2vNwhqZRMYxtEpr7dtlq4mzqkDAfBgNVHSMEGDAWgBSDOyM4U0iy
-SsNe3WvVASRLsQtxHTANBgkqhkiG9w0BAQsFAAOCAgEAU8bt/AX1kj+Q8PKoAsbn
-0buTYm3qqFevhIHzTWeznKIPDCbV1HOG34PigL9TjjxfqcBNO1BVjTDmdHMKzvAZ
-CwB5tcYX6jXGaus2F0UmrT2ORmWn5+d/G34RC2DsEnjKbHU3n76SFQw1A9KM+Smv
-wy6EiSCoJXmq89dxeTpdVblXOG4Vm0VImQr9zPwM878YZToKsGpdcBsYAE2I24nI
-WlxShImP1ph8HUZ1iFheoNVF+MS7m3KgskYcyADi3rtQyLfhXSo1GLQJBLbzBYjE
-3amsZRRXIJOIxH6EO7rrwyCAzla22vaysDhLJEjzouCdPK5Xs+EVPjULo/QXxRmP
-L2Y5ZvxE3KetkobKtmm92lD8fHK8gC7SL9PuDWj3A5r4FuzWa99oI+zUa4ebBs90
-G26m7dFtkgARRYL9eSb9NIa7tw5BBFsaDVwE1Nfj5EgFx+J6buD9vATAkQpAIn4y
-S2Fo3O4kXTwTXziVJAzraUgO3buwSF5rnAnbOO+tS7BpQga056lZzSJujvFRw1MS
-FOR/SkyWEt25o0i8lOK3M/NSZdugwDZzBLgDVcNsvxQw+so9V1eXpcFDy/KYz/3G
-oEc5PErdznByZe52vRKK76Cox3ysC36oI6wznzMawJGyBruITKSIDCV0DBLmRxkm
-291pYW8PMTNHFk6rCEsDgfU=
------END CERTIFICATE-----
