--- agent/yubiagent/testdata/Unittest_Authentication_9a.crt
@@ -1,29 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIFCjCCAvKgAwIBAgIBATANBgkqhkiG9w0BAQsFADAlMSMwIQYDVQQDExpVbml0
-dGVzdCBBdXRoZW50aWNhdGlvbiA5YTAeFw0yMjA4MjkxNzE4MzdaFw0yNDAyMjkx
-NzE4MzZaMCUxIzAhBgNVBAMTGlVuaXR0ZXN0IEF1dGhlbnRpY2F0aW9uIDlhMIIC
-IjANBgkqhkiG9w0BAQEFAAOCAg8AMIICCgKCAgEAvfiBX+2zMqSY0bFdrBbz/7At
-v6ZiOhtFlcVBr5Mh/FPSPBWYebKNOSmvdkOMBBZBhEjbGzrrgsP3CIMpYH899dDq
-U0DQqz0vuaR+1BrzUfAKv806BojzAyXS8CV2dBII6jWH8g93ocLfPJ6tivAhBkh7
-SpCHXZe6+whhNo69ohxd2bCYhEcJ/jp/N88B7HmJKMeZpmuf/vnmV9HQaxnZgQ0w
-lOqZFHL+VT678Xnv2LP3bYjH4bpjMxRRUf+Fe+7rhT0vKHYI7baXC/BBiCCo+uAn
-Dcbvj+IzZomDqlutII/9LxyccP/lLDZG9SGbW1DDtQV2mZzlo7Q9xbtlYN71Fhck
-0vEkXca8BZo8Y3wKI2sqk58zwREDy1B1LHlc7jD29KcM/cUazIUNJa8zVhh1fj4I
-zdC6qrjXsrBRALjW7N87+sgStXoehKPlz1Run5ym+eb/Cppe8U07AW6ceStafqQt
-oTSO7P3+WY/iW5G8gxFdrf4Okg+OIDn7CIUJW8ECjnWUPg6FLGIDMiZq43WzczSs
-SRgvft+90H7KLPxX3dgTcteGjO83NhJoVKE4UImBwm3bnEb8FLvSgVSS+tw6aOCE
-YqfBoo/4mWzyuXbKOpVP3V19jald8wMwkFYh12ydGqKrtkPmE4H/ZY1aloJcoxS8
-8oGKq/M6hh+HTOzQHFUCAwEAAaNFMEMwDgYDVR0PAQH/BAQDAgEGMBIGA1UdEwEB
-/wQIMAYBAf8CAQAwHQYDVR0OBBYEFDVV3wEUnc7ZPbJHZ0wv1N3CrKs2MA0GCSqG
-SIb3DQEBCwUAA4ICAQA7aNgeZc++M49/HZuqF6fRnMM4O2QsEy79llF9ytDysope
-N7REMwl6eGW5KV+5e+0shrgahySLrAldmSfNEF/lWPZ/qP7lbFBaDKBN0pbAHPWf
-Um/XLhdzoURFj9c9QO/j4LiX6LiVBr/7KC2KoeYefcuec8BKIgSiTsnXcRXgG6zA
-cOxF0XQ16qQWuMne0yBY2VJABJImezjSDbitcrsudhW9+KS42BMBE9my56WayJ7s
-NbYETW+k1uTmDTveo9H3Zt8YHyqQ2b/6BvASRQ/TPV1VJec+Pyg1Z0Qrxdzp+It5
-ESV5a83+iXW5vurzU4zPAM0FIFbbjTXpNRd0FxxH26I06WxRIeErCAgxWSvg6hjU
-Vyhmr7n1N3mhzr34SUPs4Fzqpysi+gDU3z8HzE842UwLoJn/MoAA+RQXKLY5jkIz
-pEWKnR9A47UXJmzIbFhh7QdIrkzfjVpaUpLTToDXjCYSApdDv18xZXn1U8/bFh7V
-2omXQpbWQiYHofDHVk1vMf61lAEGAWzzfZp6vknemQIPLSdmL3YGjRPJfVqWyBXn
-GcuvLB6k+c4zibc77k6c1JSZ+shIctq/GgdveRiWNRJpuUtT7rL7P/wDJ9egF9EH
-DT7WPdtY5HVQADl2zOjK1HoArOO+gicpQtbqfO38Cj9MrtNw5XhhegmQwBFsfw==
------END CERTIFICATE-----
