--- agent/yubiagent/testdata/Unittest_Authentication_9a_attest.crt
@@ -1,25 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIETDCCAzSgAwIBAgIRAKBNpl0a5egp/3IHviowVxswDQYJKoZIhvcNAQELBQAw
-IjEgMB4GA1UEAxMXVW5pdHRlc3QgQXR0ZXN0YXRpb24gZjkwIBcNMjIwODI5MTcx
-ODM3WhgPMjEwMjA4MjkxNzE4MzZaMCwxKjAoBgNVBAMTIVVuaXR0ZXN0IEF1dGhl
-bnRpY2F0aW9uIDlhIGF0dGVzdDCCAiIwDQYJKoZIhvcNAQEBBQADggIPADCCAgoC
-ggIBAL34gV/tszKkmNGxXawW8/+wLb+mYjobRZXFQa+TIfxT0jwVmHmyjTkpr3ZD
-jAQWQYRI2xs664LD9wiDKWB/PfXQ6lNA0Ks9L7mkftQa81HwCr/NOgaI8wMl0vAl
-dnQSCOo1h/IPd6HC3zyerYrwIQZIe0qQh12XuvsIYTaOvaIcXdmwmIRHCf46fzfP
-Aex5iSjHmaZrn/755lfR0GsZ2YENMJTqmRRy/lU+u/F579iz922Ix+G6YzMUUVH/
-hXvu64U9Lyh2CO22lwvwQYggqPrgJw3G74/iM2aJg6pbrSCP/S8cnHD/5Sw2RvUh
-m1tQw7UFdpmc5aO0PcW7ZWDe9RYXJNLxJF3GvAWaPGN8CiNrKpOfM8ERA8tQdSx5
-XO4w9vSnDP3FGsyFDSWvM1YYdX4+CM3Quqq417KwUQC41uzfO/rIErV6HoSj5c9U
-bp+cpvnm/wqaXvFNOwFunHkrWn6kLaE0juz9/lmP4luRvIMRXa3+DpIPjiA5+wiF
-CVvBAo51lD4OhSxiAzImauN1s3M0rEkYL37fvdB+yiz8V93YE3LXhozvNzYSaFSh
-OFCJgcJt25xG/BS70oFUkvrcOmjghGKnwaKP+Jls8rl2yjqVT91dfY2pXfMDMJBW
-IddsnRqiq7ZD5hOB/2WNWpaCXKMUvPKBiqvzOoYfh0zs0BxVAgMBAAGjcTBvMA4G
-A1UdDwEB/wQEAwIDuDAdBgNVHSUEFjAUBggrBgEFBQcDAQYIKwYBBQUHAwIwHQYD
-VR0OBBYEFDVV3wEUnc7ZPbJHZ0wv1N3CrKs2MB8GA1UdIwQYMBaAFOna83CGplEx
-jG0Smvt22WribOqQMA0GCSqGSIb3DQEBCwUAA4IBAQAmBWJGf1GNBLpWOKQjjkrz
-P1ugW5oqQcTSkuwW8pUJ7UGPNUpgzyPhrC6xvyX6H5rFYbPjLvjPQrX7L2ociNIh
-/kZrEi7q6iNppLoKVcoctBHNSaaVzvG+rMNT+FKTFRcVaetyJLmRE50tS2XXAXwy
-p4+ddbMjsmrB75pptGeUXg0DQh6TKBzg9Sp58Vufj0cBx0c4uNP0yfgh8oQSTlL8
-DykUZPpFpuR+n7KXHgedadUc0pFQit3j1pmEkVol67NE2n/a9zdzE0DtqPyMCYMk
-fN752ZyDn9TMcNWNXqZzMt9voGWyAi70C3mke40L6/XYKBsK+qlYZO8QiW9EMg+2
------END CERTIFICATE-----
