--- agent/yubiagent/testdata/Unittest_Authentication_9e.crt
@@ -1,29 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIFCjCCAvKgAwIBAgIBATANBgkqhkiG9w0BAQsFADAlMSMwIQYDVQQDExpVbml0
-dGVzdCBBdXRoZW50aWNhdGlvbiA5ZTAeFw0yMjA4MjkxNzE4MzlaFw0yNDAyMjkx
-NzE4MzdaMCUxIzAhBgNVBAMTGlVuaXR0ZXN0IEF1dGhlbnRpY2F0aW9uIDllMIIC
-IjANBgkqhkiG9w0BAQEFAAOCAg8AMIICCgKCAgEA5W9jpZfJ7sSF7Vm6ZSsxFATr
-kcga4YMFkHygFipwaqz2XZOdfMQO4drK3RtIzrzo4hQognNcN+mnhnYMc+Khbtze
-bFasTmVf/w70GIuryquxBtvR2RcwZWadzhByHWV+6hAUKcuE9PcMAcejX3c1Ucds
-X+jv/uCimK351Bb00cfITj8+n2YjeYd5q81MfGc6oEIOmxpEFpeDd6c5wfGhNSPO
-RMtXgObQ3WEdZV1PavL7Q48ci9F+DXLmjkwoaSYEff1xTl9QlE2b4014mV5GDxx+
-nnCATDyOyIWyB8UKgZBUAnuyFDAgmwTzVIbgQUmqXr9JEDWqQIHN0D+/deHHoAzW
-Ks7VqGjaV4mHNOfNBOyVQsY7qFemGJphLXO1y7BHshgh81y4E4DKw7jPfw+o+Y8C
-hy3t8tBsD7C3SSycCUEBPGsVbaXI3DeiB3vQpCdnkX4lq4Q3JNEwqRwt4Y02YlBW
-LAQPIyNwvoUHu2Vkc7t1Aa8nihj7ZEd4sDeZKkOGFtD1CImdbi3puUbuuy4970sl
-903Hv//c95HkHEfIlas5qIh16F4/YFB3q54qYkn4dalvxWjsnz2Myr0iu1lN49UH
-+aCRuFUw0npKop8CQe1AL5b2HXUhH5WNUBi+X4imHVLIVneuKLYK6FLWw2coF2uS
-iO8R4BVZTRxebH9+UaUCAwEAAaNFMEMwDgYDVR0PAQH/BAQDAgEGMBIGA1UdEwEB
-/wQIMAYBAf8CAQAwHQYDVR0OBBYEFN3HoMTlxzY8qRPOqwAwkyuHZTTsMA0GCSqG
-SIb3DQEBCwUAA4ICAQABQQEqoOjJyNJMYA7M3ptQK5lVch3uGr5FrcKvqJikJyQi
-jpWK6xsb222CeOhxotRmh0hYSYzSxPaAd5QasNRfJn/avXR72Kph8qbT46ncqnOh
-o89c5Mn9Bv0+b5LokaWFFkZ+MLx31PzO1TGHhOqkSH17DDoLxLpKilXfJ7gFFM05
-CTqjMrYx5PUGX/8JH3sTfkwz0lKS06GJO+VGamVzkhbvO6TPGzUXAeZfgp39HeRA
-snCtYJmVAnh2s+k/g+DMqtT09VAyxTbDA/IC1uDrga4zauAXhORbbr8/PsqrD2LH
-8Lw20jqQRFXwElY81n9D7TqJxo0dQc+sC8a/WRfQt6xCADgL45Q2RrwnfM5xZT+W
-z3vafzez0yOE0SoeuA3elX9h1ogzjEUwBZy0DGJhbjHWVDQ5cUJwKxIJcL3FQP71
-b+TaDcFaQEeXyz8AQzUPvkOcTJDhl4jKQiIOHpBdiSES35NU+ROYvTUppsh0PHg9
-meUF87WUCiMK4C9Ul+HPQ5mgjaO7xUDVgwfdf+L1o2CbW6sP9SMcPv64gnrHAxVE
-txvFzY/oUt778Ye6I0ITTDYF7cLlmAopBAn8RSg3UpRSyOZLn58VTk86y+DJb1wx
-IUgscDN1Z2T6EhiDMB2xYBaR6ONm9pl+vyJ0wDpxT9bWcc/Xvbvu6FmP9M2Zxg==
------END CERTIFICATE-----
