--- agent/yubiagent/testdata/Unittest_Authentication_9e_attest.crt
@@ -1,25 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIESzCCAzOgAwIBAgIQN6XeBO794CZWj7XDSfkVRDANBgkqhkiG9w0BAQsFADAi
-MSAwHgYDVQQDExdVbml0dGVzdCBBdHRlc3RhdGlvbiBmOTAgFw0yMjA4MjkxNzE4
-MzlaGA8yMTAyMDgyOTE3MTgzNlowLDEqMCgGA1UEAxMhVW5pdHRlc3QgQXV0aGVu
-dGljYXRpb24gOWUgYXR0ZXN0MIICIjANBgkqhkiG9w0BAQEFAAOCAg8AMIICCgKC
-AgEA5W9jpZfJ7sSF7Vm6ZSsxFATrkcga4YMFkHygFipwaqz2XZOdfMQO4drK3RtI
-zrzo4hQognNcN+mnhnYMc+KhbtzebFasTmVf/w70GIuryquxBtvR2RcwZWadzhBy
-HWV+6hAUKcuE9PcMAcejX3c1UcdsX+jv/uCimK351Bb00cfITj8+n2YjeYd5q81M
-fGc6oEIOmxpEFpeDd6c5wfGhNSPORMtXgObQ3WEdZV1PavL7Q48ci9F+DXLmjkwo
-aSYEff1xTl9QlE2b4014mV5GDxx+nnCATDyOyIWyB8UKgZBUAnuyFDAgmwTzVIbg
-QUmqXr9JEDWqQIHN0D+/deHHoAzWKs7VqGjaV4mHNOfNBOyVQsY7qFemGJphLXO1
-y7BHshgh81y4E4DKw7jPfw+o+Y8Chy3t8tBsD7C3SSycCUEBPGsVbaXI3DeiB3vQ
-pCdnkX4lq4Q3JNEwqRwt4Y02YlBWLAQPIyNwvoUHu2Vkc7t1Aa8nihj7ZEd4sDeZ
-KkOGFtD1CImdbi3puUbuuy4970sl903Hv//c95HkHEfIlas5qIh16F4/YFB3q54q
-Ykn4dalvxWjsnz2Myr0iu1lN49UH+aCRuFUw0npKop8CQe1AL5b2HXUhH5WNUBi+
-X4imHVLIVneuKLYK6FLWw2coF2uSiO8R4BVZTRxebH9+UaUCAwEAAaNxMG8wDgYD
-VR0PAQH/BAQDAgO4MB0GA1UdJQQWMBQGCCsGAQUFBwMBBggrBgEFBQcDAjAdBgNV
-HQ4EFgQU3cegxOXHNjypE86rADCTK4dlNOwwHwYDVR0jBBgwFoAU6drzcIamUTGM
-bRKa+3bZauJs6pAwDQYJKoZIhvcNAQELBQADggEBAGXj2o9tNFeHt1yxIFKV6p5m
-8NkAlmcLKNFN56JuDgJY6pzOl4bYkX7n8GBtPjRhEZgOoFOQ0PVl6qp2spOXj76m
-1yUbCz+hA89k3RBX5tzSu67UxeJub+n35dFuvGC585IgD8nA3mcZSujqsp1bYG0X
-yJrTFxcREFRyCtz78Tip17JoRlRb3DcrDBvL2fI+jvunO8hoXu+RZgy3q6X14+Fs
-MDqm1s5cSYAiJqUa47cYB4nNq7nRuiWcsvSDM13Wm/QqFxhW5zc6sv667VLlKTSM
-d2EC/CSKM10eXmDNNKAxMFpDnZH+KCQbTr0RNgT9sgyOu7PkFlsAJNt7Ev7eiIs=
------END CERTIFICATE-----
