--- agent/yubiagent/testdata/Unittest_Root_CA.crt
@@ -1,29 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIE+DCCAuCgAwIBAgIBATANBgkqhkiG9w0BAQsFADAbMRkwFwYDVQQDExBVbml0
-dGVzdCBSb290IENBMCAXDTIyMDgyOTE3MTgzNloYDzIxMDQwMjI5MTcxODMzWjAb
-MRkwFwYDVQQDExBVbml0dGVzdCBSb290IENBMIICIjANBgkqhkiG9w0BAQEFAAOC
-Ag8AMIICCgKCAgEAzStwChyLXaQ4PPAiSLcRs37DLKEUypcnsWC9cvpDy0fvQOwH
-kePSe/mCMQNgratnN46dra39Oz0mxbDS+F32tqY9lmA6lRuJ4aEBu+WVba4AhKn2
-3HUlDN0ztzY0wSepuhKDq5M9nJHCYYMJku36YxUWvuwLQ7b0Q/W1/8L3Kfvmr0wi
-ZQWESoHPnS/pKQD3x4vIAvF8AmeUuRPWsTUn/kALaFNXiUQKj5dZt9GwcSz3Uy5j
-D/dzH8Uw859fniz4Xi46RVfNNQRO/73QV+UF2/7mk2SSCH/RHnud3G1gIIzaK6KB
-zieFLGrs8V2mmc9hnD6C3zlGii0yqj1Yev43mW80dPz93TxWArKxjz0XR2ng3+T4
-l7XY1+T+wVbX+aV80pSzQdQF1V9S1P3f/asbnKyc/oF4PE2AUn4hGqelTeN7dUvD
-XXRB0A7bjwH22PGY2oPFiBAQiSFLqmuL80+qC5X1syevz8xfQEEb0fvWzV4F7EBT
-vY0hVDjM0LB6AOZZFO0a24W2zUs+K6+gfblb4Bi7s2S+qaiHcfdYO6pGhlxlduQP
-GqqKhZ0ArYwrq1tNcomLr9uy6f1Dnv1mDGWx54MHzxX8ZjvMKloz1V+yrV+XHPqu
-FKWBUqTcfX5+BVAO/5Cq6fzng7nnHeVppXdLXDP6Jy5eJr36+HFrBChhwK8CAwEA
-AaNFMEMwDgYDVR0PAQH/BAQDAgEGMBIGA1UdEwEB/wQIMAYBAf8CAQAwHQYDVR0O
-BBYEFIM7IzhTSLJKw17da9UBJEuxC3EdMA0GCSqGSIb3DQEBCwUAA4ICAQA1dhAQ
-dWR6S+agLNIBG++yggJdyjIaRA4p5/M5gfkTHYcqWh8XovgxwPcNAZeToYN6f/AV
-RSo/IMnnS2wiPv85IF0mZTrUVn1Vf5LjboOP7ZLPRvPWJfF6Vo2sLKjTlwOsOfJ6
-CEWJuhljsl5kj2dByGop2yv+3hV6QIm6VOTIKzrzqbzvbAoc1qf0I6cxFpkZW06f
-l3q3h0Wz4PvI7K/eAI4L515mTVXz95W4RTV0CFUnEKWq15W/Y+OA7xFCqeslAmRv
-nYsAOZkZCl+kezD6JFt7J3BU+ypIMSakbenj3OegliBgs8VKSpwsaGX2RIN0s7a/
-kWrt/jxqg6zX/MZoykzF0bmA6Xfl3Ed6HRhYTNxVx/5eojmbN8D8v6PBa0CjlHKV
-dy1y37Tc5kaAKZ9fF45Y0X0Jky0a4cF1HKjjRjB6QgbOQyi8isSSM1WFBc++mP13
-KEq0gxps3bGdrMeDfckyB3zIN8+V5IR08T9aG0HFCTLbc5ghMSWCLPRb+AiXwlFt
-h1ne7l46uPQHPQlMHVbnQliqNyED8QFOpHsEodzqk+bYhMJxlRpUlf3TKz3WBFqt
-F9Uer3Zt34RdlTvgQH7QAe83hoyMZLHru73D0MYubxFRF/8N4D0mYtAZwzlYXO8R
-MNo+gPu+gLlh02TxKMNeVH/kF4KAGARJdwgidw==
------END CERTIFICATE-----
